CrypNotes/requirements.txt
fastapi
uvicorn
pydantic
pycryptodome
jinja2
python-multipart
